Apply naming convention for css classes, avoid collision with fluentui classes, and improve readability + update tests
