+++ src/web/components/SideBar/SideBar.css
@@ -1,0 +1,72 @@
+.ftk-sidebar {
+    min-height: 100dvh;
+    display: flex;
+    flex-direction: column;
+    position: relative;
+    transition: width 0.2s ease;
+    padding: 40px 16px 16px 16px;
+    margin-top: 0;
+
+    &.ftk-collapsed {
+        width: 80px;
+    }
+
+    &.ftk-expanded {
+        width: 216px;
+    }
+
+    .ftk-sidebar-toggle {
+        position: absolute;
+        top: 10px;
+        right: 8px;
+    }
+
+    .ftk-sidebar-menu {
+        display: flex;
+        flex-direction: column;
+    }
+
+    .ftk-sidebar-item {
+        display: flex;
+        align-items: center;
+        padding: 10px 12px;
+        color: var(--colorNeutralForeground1);
+        text-decoration: none;
+        border-radius: 6px;
+        position: relative;
+        transition: background-color 0.2s ease, color 0.2s ease;
+
+        &:hover {
+            background-color: var(--colorNeutralBackground2Hover);
+            color: var(--colorNeutralForeground2BrandHover);
+        }
+
+        .ftk-sidebar-icon {
+            margin-right: 10px;
+            transition: color 0.1s ease;
+            color: var(--colorNeutralForeground2);
+
+        }
+
+        &:hover .ftk-sidebar-icon {
+            color: var(--colorNeutralForeground2BrandHover);
+        }
+
+        &::before {
+            content: "";
+            position: absolute;
+            left: 2px;
+            top: 22px;
+            transform: translateY(-50%);
+            width: 4px;
+            height: 18px;
+            background-color: transparent;
+            border-radius: 2px;
+            transition: background-color 0.1s ease;
+        }
+
+        &:hover::before {
+            background-color: var(--colorNeutralForeground3BrandHover);
+        }
+    }
+}
