+++ src/web/tsconfig.app.json
@@ -20,5 +20,5 @@
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true
   },
-  "include": ["pages", "pages/App.tsx"]
+  "include": ["pages", "pages/App.tsx", "components/__tests__/SideBar.test.tsx"]
 }
